Made bap lib propogate deps needed at linktime
OCaml does a funny thing where it assumes the libraries a package
depends on will again be on the search path during a link against this
library. To compensate for this, I've moved a bunch of libraries into
propagated deps.

This still isn't quite right - it includes some transdeps, like ncurses,
because tracking down all the other OCaml packages that didn't realize
this propagation requirement would take a while. It's not suitable for
release like this, but at least packages depending on bap only need to
add "bap" to their build inputs.